fix: 수정 사안 반영 #24

+++ src/main/java/heeyoung/hee/domain/Assignment/service/AssignmentService.java
@@ -9,9 +9,11 @@ import heeyoung.hee.domain.User.entity.User;
 import heeyoung.hee.domain.User.repository.UserRepository;
 import heeyoung.hee.global.exception.ErrorCode;
 import heeyoung.hee.global.exception.RestApiException;
+import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.data.domain.Pageable;
 
 import java.util.Date;
 import java.util.List;
@@ -24,9 +26,10 @@ public class AssignmentService {
     private final AssignmentRepository assignmentRepository;
 
     // 과제 제출
+    @Transactional
     public AssignmentResponseDto submitAssignment(AssignmentSubmitDTO dto, Long userId) {
 
-        // 유저 등록
+        // 유저 조회
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RestApiException(ErrorCode.USER_NOT_FOUND));
 
@@ -47,10 +50,16 @@ public class AssignmentService {
     }
 
     // 과제 수정
-    public AssignmentResponseDto updateAssignment(AssignmentUpdateDTO dto, Long assignmentId) {
+    @Transactional(readOnly = true)
+    public AssignmentResponseDto updateAssignment(User user, AssignmentUpdateDTO dto, Long assignmentId) {
 
         Assignment assignment = assignmentRepository.findById(assignmentId)
                 .orElseThrow(()->new RestApiException(ErrorCode.ASSIGNMENT_NOT_FOUND));
+
+        // 유저 검증
+        if (!assignment.getUser().getId().equals(user.getId())) {
+            throw new RestApiException(ErrorCode.USER_NOT_MATCH);
+        }
 
         // Assignment 갱신
         Assignment updatedAssignment = assignment.update(dto.getTitle(), dto.getContent(), dto.getLink());
@@ -63,27 +72,25 @@ public class AssignmentService {
     }
 
     // 과제 삭제
-    public void deleteAssignment(Long assignmentId) {
+    @Transactional(readOnly = true)
+    public void deleteAssignment(User user, Long assignmentId) {
 
         Assignment assignment = assignmentRepository.findById(assignmentId)
                 .orElseThrow(()->new RestApiException(ErrorCode.ASSIGNMENT_NOT_FOUND));
+
+        // 유저 검증
+        if (!assignment.getUser().getId().equals(user.getId())) {
+            throw new RestApiException(ErrorCode.USER_NOT_MATCH);
+        }
 
         assignmentRepository.delete(assignment);
     }
 
     // 과제 전체 조회
-    public List<AssignmentResponseDto> findAllAssignments(String option, String direction) {
-        List<Assignment> assignments = List.of();
+    @Transactional(readOnly = true)
+    public List<AssignmentResponseDto> findAllAssignments(Pageable pageable) {
 
-        // 오름차순
-        if (direction.equalsIgnoreCase("asc")) {
-            assignments = assignmentRepository.findAll(Sort.by(Sort.Direction.ASC, option));
-        }
-
-        // 내림차순
-        else if (direction.equalsIgnoreCase("desc")) {
-            assignments = assignmentRepository.findAll(Sort.by(Sort.Direction.DESC, option));
-        }
+        List<Assignment> assignments = assignmentRepository.findAll(pageable).getContent();
 
         return assignments.stream()
                 .map(AssignmentResponseDto::from)
